% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_limes.R
\name{compare_limes}
\alias{compare_limes}
\title{Compute the locality metric for LIME implementations}
\usage{
compare_limes(explanations)
}
\description{
Computes the locality metric measuring how well the explainer
models approximate the complex model for different LIME
implementations.
}
