% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare_metrics.R
\name{compare_metrics}
\alias{compare_metrics}
\title{Plot of LIME Comparison Metrics}
\usage{
compare_metrics(explanations, metrics = NULL)
}
\arguments{
\item{explanations}{explain dataframe from apply_lime}

\item{metrics}{}
}
\description{
Plots the specified comparison metrics by LIME
tunning parameters
}
