% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-feature_heatmap.R
\name{feature_heatmap}
\alias{feature_heatmap}
\title{Heatmap of LIME Selected Features}
\usage{
feature_heatmap(explanations, feature_nums = NULL)
}
\arguments{
\item{explanations}{explain data frame from the apply_lime list}

\item{feature_nums}{accepts a vector of integer values from 1 to \code{nfeatures} specified in \code{apply_lime} to specify which features selected by LIME should be included in the plot}
}
\description{
Creates a heatmap of the features selected by lime for each
of the tunning parameters
}
