% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explain_scatter.R
\name{plot_explain_scatter}
\alias{plot_explain_scatter}
\title{Visualize the explanation of interest (eoi) and LIME simulated data}
\usage{
plot_explain_scatter(
  explanation,
  bins = TRUE,
  weights = TRUE,
  alpha = 1,
  line_size = 1,
  title.opt = TRUE
)
}
\arguments{
\item{explanation}{Rows from the LIME explanations associated 
with one prediction of interest}

\item{bins}{Should lines indicating the bins used by LIME be included? 
Only applicable is sim_method is equal to "quantile_bins" or 
"equal_bins". (Default is TRUE.)}

\item{weights}{Should the size of the points represent the weight 
assigned by LIME? (Default is TRUE.)}

\item{alpha}{Value to use for alpha blending of the points}

\item{line_size}{Size of the lines used for representing the explainer model}

\item{title.opt}{Should a title be included that lists the 
simulation method and Gower exponent? (Default is TRUE.)}
}
\description{
Visualize the explanation of interest (eoi) and LIME simulated data
}
\examples{

# Prepare training and testing data
x_train = sine_data_train[c("x1", "x2", "x3")]
y_train = factor(sine_data_train$y)
x_test = sine_data_test[1:5, c("x1", "x2", "x3")]

# Fit a random forest model
rf <- randomForest::randomForest(x = x_train, y = y_train)
                       
# Run apply_lime
res <- apply_lime(train = x_train, 
                  test = x_test, 
                  model = rf,
                  label = "1",
                  n_features = 2,
                  sim_method = c('quantile_bins',
                                 'kernel_density'),
                  nbins = 2:3, 
                  return_perms = TRUE)
 
# Extract the rows associtaed with an explanation 
# of interest (the first observation in the test data) 
eoi <- res$explain[1:2,]

# Plot the explanation of interest
plot_explain_scatter(eoi)
}
