% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sine_data.R
\docType{data}
\name{sine_data}
\alias{sine_data}
\title{Sine Classifier Data}
\format{A data.frame.}
\usage{
sine_data
}
\description{
A dataset that contains...
}
\details{
The variables in the dataset are as follows.
\tabular{ll}{
 \code{x} \tab \cr
 \code{x2} \tab \cr
 \code{x3} \tab \cr
 \code{y} \tab }
}
\keyword{datasets}
