% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-apply_lime.R
\name{apply_lime}
\alias{apply_lime}
\title{Apply Different Implementations of LIME}
\usage{
apply_lime(train, test, model, sim_method, nbins = 4, label, n_features,
  n_permutations = 5000, feature_select = "auto", dist_fun = "gower",
  kernel_width = NULL, gower_pow = 1, all_fs = FALSE,
  return_perms = FALSE, label_fs = NULL, seed = NULL)
}
\arguments{
\item{train}{Dataframe of training data features.}

\item{test}{Dataframe of testing data features.}

\item{model}{Complex model to be explained.}

\item{sim_method}{Vector of methods to use for creating the simulated
data. Options are 'quantile_bins', 'equal_bins',
'kernel_density', and 'normal_approx'.}

\item{nbins}{Vector of number of bins to use with bin based
simulation methods.}

\item{label}{Response category to use in the explanations. Current
implementation only accepts 1 label.}

\item{n_features}{Number of features to return in the explanations.}

\item{n_permutations}{Number of permutations to use when simulating
data for each explanation. Default is 5000.}

\item{feature_select}{Feature selection method. Options are 'auto',
'none', 'forward_selection', 'highest_weights', 'lasso_path',
and 'tree'.}

\item{dist_fun}{Distance function to use when computing weights for
the simulated data. Default is 'gower'. Otherwise,
\code{stats::dist()} will be used.}

\item{kernel_width}{Kernel width to use if \code{dist_fun} is not
'gower'.}

\item{gower_pow}{Numeric vector of powers to use when computing 
the Gower distance.}

\item{all_fs}{Indicates whether all feature selection methods
should be applied for an implementation of LIME to see how
the features selected varies within a LIME implemenation.
Note that the LIME results returned will correspond to the 
method specified in the \code{feature_selection} option.}

\item{return_perms}{Should the simulated dataset (permutations) be 
returned for all of the observations in the test datatset and
LIME implementations? Default is FALSE.}

\item{seed}{Number to be used as a seed (if desired).}
}
\description{
Applies LIME with the specified tunning parameter options.
}
\examples{

# Prepare training and testing data
x_train = sine_data_train[c("x1", "x2", "x3")]
y_train = factor(sine_data_train$y)
x_test = sine_data_test[1:5, c("x1", "x2", "x3")]

# Fit a random forest model
rf <- randomForest::randomForest(x = x_train, y = y_train) 

# Run apply_lime
res <- apply_lime(train = x_train, 
                  test = x_test, 
                  model = rf,
                  label = "1",
                  n_features = 2,
                  sim_method = c('quantile_bins',
                                 'kernel_density'),
                  nbins = 2:3)
}
