% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_heatmap.R
\name{feature_heatmap}
\alias{feature_heatmap}
\title{Heatmap of LIME Selected Features}
\usage{
feature_heatmap(explanations, feature_nums = NULL)
}
\description{
Creates a heatmap of the features selected by lime for each
of the tunning parameters
}
