% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-feature_heatmap.R
\name{feature_heatmap}
\alias{feature_heatmap}
\title{Heatmap of LIME Selected Features}
\usage{
feature_heatmap(explanations, feature_nums = NULL)
}
\arguments{
\item{explanations}{Explain dataframe from the list returned by apply_lime.}

\item{feature_nums}{A vector of integer values from 1 to \code{nfeatures} (specified in \code{apply_lime}) to determine which features selected by LIME should be included in the plot.}
}
\description{
Creates a heatmap of the features selected by lime for all
observations in the test set across all of the different
LIME implementations.
}
\examples{

# Create Random Forest model on the sine data
rfsine <- caret::train(x = sine_data_train[c("x1", "x2", "x3")],
                       y = sine_data_train$y,
                       method = "rf")

# Apply lime several implementations of LIME
sine_lime_explain <-
   apply_lime(train = sine_data_train[c("x1", "x2", "x3")],
              test = sine_data_test[c("x1", "x2", "x3")],
              model = rfsine,
              label = "1",
              n_features = 2,
              sim_method = c('quantile_bins', 'kernel_density'),
              nbins = c(3, 4),
              seed = 20190914)

# Plot heatmap of selected features across LIME implementations
feature_heatmap(sine_lime_explain$explain)

# Return a heatmap with only the features selected first by LIME
feature_heatmap(sine_lime_explain$explain, feature_num = 1)
}
