% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plot_compare.R
\name{plot_compare}
\alias{plot_compare}
\title{Plot of LIME Comparison Metrics}
\usage{
plot_compare(explanations, metrics = NULL)
}
\description{
Plots the specified comparison metrics by LIME
tunning parameters
}
